Fix Meziantou.Analyzer warnings across all projects
- Remove unused methods (Be32, ReadBe64, DecodeNormal, DecodeWithMatchByte, etc.)
- Fix captured variable disposed in outer scope (LibraryApiTests)
- Remove optional params hidden by overload (LzmaStream.Create)
- Remove unused constants/variables (GroupDataSize, H2Size, H2Padding, written, usedIndex)
- Remove unused parameter (blocksPerChunk in WiiPartitionExtractor)
- Fix possible overflow in unchecked context (Decoder.Fast.cs)
- Add StructLayout to structs (MA0008)
- Add param name to ArgumentException (MA0015)
- Suppress style rules (MA0002, MA0005, MA0006, MA0011, MA0016, MA0048, MA0074)

+++ RVZSharp/Compression/Lzma/Decoder.Fast.cs
@@ -475,7 +475,7 @@ public partial class Decoder
         // logical-shift result from a target of 0, so both formulas collapse into one
         // branchless expression selected by the same mask used above.
         int target = (int)(
-            (RangeCoder.BitDecoder.K_BIT_MODEL_TOTAL & ~mask) | (KBitModelOffsetFast & mask)
+            (RangeCoder.BitDecoder.K_BIT_MODEL_TOTAL & ~mask) | ((uint)KBitModelOffsetFast & mask)
         );
         *probSlot = (ushort)((int)prob + ((target - (int)prob) >> KNumMoveBitsFast));
 
@@ -521,7 +521,7 @@ public partial class Decoder
     private static unsafe void UpdateProbFast(ushort* probSlot, uint prob, uint mask)
     {
         int target = (int)(
-            (RangeCoder.BitDecoder.K_BIT_MODEL_TOTAL & ~mask) | (KBitModelOffsetFast & mask)
+            (RangeCoder.BitDecoder.K_BIT_MODEL_TOTAL & ~mask) | ((uint)KBitModelOffsetFast & mask)
         );
         *probSlot = (ushort)((int)prob + ((target - (int)prob) >> KNumMoveBitsFast));
     }
